+++ system/src/templates.mjs
@@ -78,9 +78,12 @@ export default async function preloadHandlebarsTemplates() {
 		"systems/fallout/templates/actor/settlement/partials/structures.hbs",
 		"systems/fallout/templates/actor/settlement/tabs/status.hbs",
 		"systems/fallout/templates/actor/settlement/tabs/stockpile.hbs",
-		"systems/fallout/templates/item/parts/item-effects.hbs",
-		"systems/fallout/templates/item/parts/item-header.hbs",
-		"systems/fallout/templates/item/tabs/effects.hbs",
+		"systems/fallout/templates/item/_shared-partials/description-tab.hbs",
+		"systems/fallout/templates/item/_shared-partials/effects-tab.hbs",
+		"systems/fallout/templates/item/_shared-partials/header.hbs",
+		"systems/fallout/templates/item/weapon/tabs/attributes.hbs",
+		"systems/fallout/templates/item/weapon/tabs/mods.hbs",
+		"systems/fallout/templates/item/weapon/tabs/qualities-and-effects.hbs",
 	];
 
 	const paths = {};