+++ CHANGELOG.md
@@ -1,10 +1,16 @@
 # v11.5.0
 
 ## Enhancements
+- [#222] Add Butchery section to Creature sheet
 - [#227] Added 10mm Auto Pistol, Buzz-Saw, Flamer, Laser Emitter and Pincer robot arm attachment weapons
 - [#229] Add ammo items for thrown and explosive weapons
+- [#232] When adding a new item to a actor sheet with the add controls, auto-open the newly created item for editing
 - [#233] Make weapon ammo a select option from available ammo types in compendiums
 - [#234] Add draggable item shortcut for configured ammo on weapon sheet
+- [#238] Tidy up the caps display on NPC sheets
+- [#239] Add location to specify Wealth level on NPC sheet
+- [#240] Make it easy to roll Caps owned from an NPC's Wealth rating
+- [#241] Move weapon Qualities/Effects to their own tab
 
 ## Bugfixes
 - [#225] Weapon Qualities and Damage Effects not showing number in chat